+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Module bare specifier rewriting Babel plugin now works on Windows, does not rewrite fully qualified URLs, and will follow the "module" or "jsnext:main" fields when a package.json uses them instead of "main".
 * Add exponentiation, async/await, and async generator support when compiling to ES5.
+* Fix import specifier rewriting when importing a dependency from a top-level module.
 
 ## [0.24.0](https://github.com/PolymerLabs/polyserve/tree/0.24.0) (2018-02-23)